Remove subgroup for AAI vehicles compatibility
Remove the custom subgroup for airplanes, because it messes with the item grouping when AAI programmable vehicles is present.

As a result, the non-AI versions of the aircraft will show up on the logistics tab whereas the AI versions end up on the vehicles tab, which is quite confusing.

Without the subgroup, all aircraft versions show up on the vehicle tab.